+++ examples/one_bot.py
@@ -9,16 +9,13 @@ from two_d_game_ai.world.world import World
 configure_logger()
 
 # Create a World
-the_world = World(400)
+the_world = World(20)
 
 # Add a Bot to the World...
-the_world.add_entity(Bot(position_from_sequence=(20, 20)))
+the_world.add_entity(Bot(position_from_sequence=(2, 2)))
 
 # Create a View of the World
-view = View(
-    world=the_world,
-    scale_factor=0.7,
-)
+view = View(world=the_world)
 
 while view.running:
     view.handle_inputs()