+++ two_d_game_ai/view/view.py
@@ -33,8 +33,8 @@ class View:
     """The `World` to be rendered."""
     world_renderer_name: InitVar[str] = "UNNAMED WORLD_RENDERER"
     """Passed to `self.world_renderer`."""
-    scale_factor: float = 40
-    """Scale factor applied to the `World` render."""
+    scale_factor: float = 16
+    """Scale factor applied to the `World` render. Display units per world unit."""
     show_debug_while_unpaused: bool = False
 
     running: bool = field(init=False)